Support for JS features in markup loaded via XHR.
Adding a X-JavascriptFeatures-Init header to an XHR response that contains a space-separated list of feature names will cause those features to be initialised when the request completes.

All `global_init` functions will also be called when a jQuery XHR request completes.

If the jQuery.ajax `context` setting was used, that context will be passed to the init and global_init functions to prevent unintended changes to the rest of the page.